refactor: improvements in controller validations

+++ src/app/controllers/CategoryController.js
@@ -1,4 +1,5 @@
-const CategoriesRepository = require('../Repositories/CategoriesRepository');
+const CategoriesRepository = require('../repositories/CategoriesRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class CategoryController {
   async index(req, res) {
@@ -20,6 +21,10 @@ class CategoryController {
     const { id } = req.params;
     const { name } = req.body;
 
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ Error: 'Category id invalid' });
+    }
+
     if (!name) {
       return res.status(400).json({ Error: 'Name is required' });
     }
@@ -30,6 +35,10 @@ class CategoryController {
 
   async delete(req, res) {
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ Error: 'Contact id invalid' });
+    }
 
     await CategoriesRepository.delete(id);
     res.sendStatus(204);
